+++ src/froGH/froGH/View-Display/CameraReport_Attributes.cs
@@ -1,0 +1,35 @@
+﻿using Grasshopper.GUI;
+using Grasshopper.GUI.Canvas;
+using Grasshopper.Kernel.Attributes;
+
+namespace froGH
+{
+    internal class CameraReport_Attributes : GH_ComponentAttributes
+    {
+        public CameraReport_Attributes(CameraReport owner) : base(owner)
+        { }
+
+        protected override void Layout()
+        {
+            base.Layout();
+        }
+
+        public override GH_ObjectResponse RespondToMouseDoubleClick(GH_Canvas sender, GH_CanvasMouseEvent e)
+        {
+            bool flag = this.ContentBox.Contains(e.CanvasLocation);
+            GH_ObjectResponse result;
+            if (flag)
+            {
+                CameraReport component = (CameraReport)this.Owner;
+                component.ExpireSolution(true);
+                result = GH_ObjectResponse.Handled;// 3;
+
+            }
+            else
+            {
+                result = 0;
+            }
+            return result;
+        }
+    }
+}